+++ test/prompts/processors/json.test.ts
@@ -1,5 +1,5 @@
 import * as fs from 'fs';
-import { processJsonFile } from '../src/prompts/processors/json';
+import { processJsonFile } from '../../../src/prompts/processors/json';
 
 jest.mock('fs');
