+++ test/prompts/processors/jsonl.test.ts
@@ -1,5 +1,5 @@
 import * as fs from 'fs';
-import { processJsonlFile } from '../src/prompts/processors/jsonl';
+import { processJsonlFile } from '../../../src/prompts/processors/jsonl';
 
 jest.mock('fs');
