+++ test/prompts/processors/markdown.test.ts
@@ -1,5 +1,5 @@
 import * as fs from 'fs';
-import { processMarkdownFile } from '../src/prompts/processors/markdown';
+import { processMarkdownFile } from '../../../src/prompts/processors/markdown';
 
 jest.mock('fs');
