+++ test/prompts/processors/python.utils.test.ts
@@ -1,13 +1,16 @@
 import { PythonShell } from 'python-shell';
-import logger from '../src/logger';
-import { pythonPromptFunction, pythonPromptFunctionLegacy } from '../src/prompts/processors/python';
-import { runPython } from '../src/python/pythonUtils';
-import type { ApiProvider } from '../src/types';
+import logger from '../../../src/logger';
+import {
+  pythonPromptFunction,
+  pythonPromptFunctionLegacy,
+} from '../../../src/prompts/processors/python';
+import { runPython } from '../../../src/python/pythonUtils';
+import type { ApiProvider } from '../../../src/types';
 
 jest.mock('fs');
 jest.mock('python-shell');
-jest.mock('../src/logger');
-jest.mock('../src/python/pythonUtils');
+jest.mock('../../../src/logger');
+jest.mock('../../../src/python/pythonUtils');
 
 describe('pythonPromptFunction', () => {
   interface PythonContext {
