+++ test/prompts/processors/string.test.ts
@@ -1,5 +1,5 @@
-import { processString } from '../src/prompts/processors/string';
-import type { Prompt } from '../src/types';
+import { processString } from '../../../src/prompts/processors/string';
+import type { Prompt } from '../../../src/types';
 
 describe('processString', () => {
   it('should process a valid string prompt without a label', () => {
