+++ test/prompts/processors/text.test.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
-import { PROMPT_DELIMITER } from '../src/prompts/constants';
-import { processTxtFile } from '../src/prompts/processors/text';
+import { PROMPT_DELIMITER } from '../../../src/prompts/constants';
+import { processTxtFile } from '../../../src/prompts/processors/text';
 
 jest.mock('fs');
